Removed identifier from Confirm, PassReset and Login modules, and changed context option to token.
Previously, in the Confirm, PassReset and Login modules, it was necessary for the
developer to add the user-identifier that the verify function was meant to use
to find the user. This can now be handled by the user context module's get_by
function, so this option has now been removed.

The context option, which referred to the token context, in the Authenticate
and Remember Plugs was a bit confusing due to the word context being used
to refer to the module that now interacts with the database in Phoenix 1.3.
This option has now been changed to token and the documentation refers to
it as the token key source, rather than token context.