Made the options for endpoint and token_salt more umbrella-friendly and made sure that the key generator options are passed on to the Token module.
The endpoint and token_salt config values can now be overridden by using
keyword arguments. This is especially useful when the same config applies
to several apps, as in the case with umbrella apps.

Previously, the key generator options for signing and verifying tokens
were not being passed on from the Authenticate, or Confirm module,
to the Token module. This has been fixed, and so in these modules,
the developer can set different options for the keys.